4-general/g5-gke-ingress/artifacts/discovery/operator/app/requirements.txt
fastapi==0.115.8
uvicorn==0.17.6
kopf==1.37.4
kubernetes==32.0.0
google-cloud-dns==0.34.1
